added burnt area setup

+++ landscape.cpp
@@ -5,5 +5,5 @@ landscape::landscape(const unsigned x_size,
                      const unsigned N_trees)
     : x_size(x_size), y_size(y_size) {
     distance_map =
-        std::vector<std::vector<double>>(x_size, std::vector<double>(y_size, 0.0));
+        std::vector<std::vector<float>>(x_size, std::vector<float>(y_size, 0.0));
 }
